ref: add file_exists check in generate_module_skeletons.php to avoid creation of existing modules

+++ tools/generate_module_skeletons.php
@@ -188,6 +188,10 @@ function createModule(string $root, string $module, array $config): void
  */
 function write(string $path, string $content): void
 {
+    if (file_exists($path)) {
+        return;
+    }
+
     file_put_contents($path, $content);
 }
